requirements.txt

wheel
requests
imageio == 2.8.0
matplotlib == 3.2.2
numpy == 1.16
tensorflow-gpu == 1.14.0
keras == 2.3.1
